Refactor testing for snapshotter
Create a common base test class for snapshotter tests
